ESQL: Move more out of esql-core
This moves more things from esql-core into the esql package in our
effort to merge the two.
