+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/parser/CaseChangingCharStream.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.esql.core.parser;
+package org.elasticsearch.xpack.esql.parser;
 
 import org.antlr.v4.runtime.CharStream;
 import org.antlr.v4.runtime.misc.Interval;
